+++ tests/testthat/test-ensure.R
@@ -5,10 +5,10 @@ context("ensure")
 #' @param length numeric. The length of the random string to generate.
 #' @param alphabet character. A list of characters to draw from to create the string.
 random_string <- ensure(
-  pre = list(length %is% numeric,
+  pre = list(length %is% numeric, length(length) == 1, length > 0,
     alphabet %is% list || alphabet %is% vector,
     alphabet %contains_only% simple_string,
-    length > 0),
+    all(sapply(alphabet, nchar) == 1)),
   post = list(result %is% simple_string, nchar(result) == length),
   function(length, alphabet) {
     paste0(sample(alphabet, length, replace = TRUE), collapse = "")
@@ -46,7 +46,7 @@ describe("precondition validations", {
   })
   test_that("can have multiple errors", {
     expect_error(random_string(-10, list(1, 2, 3)),
-      "Error on alphabet %contains_only% simple_string, length > 0")
+      "Error on length > 0, alphabet %contains_only% simple_string")
   })
   test_that("default args are also checked", {
     add_default <- ensure(pre = list(x %is% numeric, y %is% numeric), post = result %is% numeric,
@@ -127,16 +127,36 @@ describe("passing conditions", {
     expect_is(rand_string, "character")
     expect_true(nchar(rand_string) == 10)
   })
+  test_that("random_string works with both pre- and post- checks and explicit formals", {
+    rand_string <- random_string(length = 10, alphabet = LETTERS)
+    expect_is(rand_string, "character")
+    expect_true(nchar(rand_string) == 10)
+  })
   test_that("add works with both pre- and post- checks", {
     add <- ensure(pre = list(x %is% numeric, y %is% numeric), post = result %is% numeric,
       function(x, y) x + y)
     expect_equal(add(4, 5), 9)
     expect_equal(add(4L, 5L), 9L)
   })
+  test_that("add works with both pre- and post- checks and explicit formals", {
+    add <- ensure(pre = list(x %is% numeric, y %is% numeric), post = result %is% numeric,
+      function(x, y) x + y)
+    expect_equal(add(x = 4, y = 5), 9)
+    expect_equal(add(x = 4L, y = 5L), 9L)
+  })
   test_that("function works with a default argument", {
     add_default <- ensure(pre = list(x %is% numeric, y %is% numeric), post = result %is% numeric,
       function(x, y = 1) x + y)
     expect_equal(add_default(4, 5), 9)
     expect_equal(add_default(4), 5)
+  })
+  test_that("function works with a default argument and explicit formals", {
+    add_default <- ensure(pre = list(x %is% numeric, y %is% numeric), post = result %is% numeric,
+      function(x, y = 1) x + y)
+    expect_equal(add_default(x = 4, y = 5), 9)
+    expect_equal(add_default(x = 4), 5)
+  })
+  test_that("can't validate twice", {
+    expect_error(ensure(pre = x %is% numeric, random_string), "already been validated")
   })
 })